Fixing items json
Error in the serialization that item class longer than 2 bytes were not serialized properly

+++ src/externalized/ItemModel.cc
@@ -153,7 +153,7 @@ void ItemModel::serializeTo(JsonObject &obj) const
 {
     obj.AddMember("itemIndex", itemIndex);
     obj.AddMember("internalName", internalName);
-    obj.AddMember("usItemClass", (uint16_t)getItemClass());
+    obj.AddMember("usItemClass", (uint32_t)getItemClass());
     obj.AddMember("ubClassIndex", getClassIndex());
     obj.AddMember("ubCursor",  getCursor());
     obj.AddMember("ubGraphicType", getGraphicType());
